@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --mode production",
-    "start": "webpack-dev-server --mode development --open"
+    "start": "webpack-dev-server --mode development"
   },
   "repository": {
     "type": "git",
@@ -21,31 +21,33 @@
   },
   "homepage": "https://github.com/OTrainFans/OTrainMap#readme",
   "devDependencies": {
+    "@types/d3-ease": "^1.0.9",
     "@types/mapbox-gl": "^0.54.5",
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.9.3",
-    "@types/react-map-gl": "^5.0.2",
-    "@types/react-virtualized": "^9.21.5",
-    "css-loader": "^3.2.0",
-    "file-loader": "^4.2.0",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-map-gl": "^5.1.1",
+    "@types/react-virtualized": "^9.21.8",
+    "css-loader": "^3.4.2",
+    "file-loader": "^4.3.0",
     "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.13.0",
-    "sass-loader": "^8.0.0",
+    "mini-css-extract-plugin": "^0.8.2",
+    "node-sass": "^4.13.1",
+    "sass-loader": "^8.0.2",
     "source-map-loader": "^0.2.4",
     "ts-loader": "^6.2.1",
-    "typescript": "^3.7.2",
-    "url-loader": "^2.2.0",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0"
+    "typescript": "^3.7.5",
+    "url-loader": "^2.3.0",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   },
   "dependencies": {
-    "mapbox-gl": "^1.5.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-map-gl": "^5.1.1",
-    "react-virtualized": "^9.21.1",
+    "d3-ease": "^1.0.6",
+    "mapbox-gl": "^1.7.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-map-gl": "^5.2.3",
+    "react-virtualized": "^9.21.2",
     "reset-css": "^5.0.1"
   }
 }
